Move save to file to export screen
