+++ core/src/main/java/com/tikz/grid/DrawType.java
@@ -6,5 +6,14 @@ public enum DrawType {
     MULTI_LINE,
     LINE,
     TEXT,
-    BEZIER
+    BEZIER;
+
+    public enum LineThickness {
+        ULTRA_THIN,
+        VERY_THIN,
+        THIN,
+        THICK,
+        VERY_THICK,
+        ULTRA_THICK;
+    }
 }
